V2 xION must have serials that have the 1st byte as 00.
Requiring that v2 xION use serials that have the first byte as 00 will make xION v2 serials unique from v1 serials. This prevents a 'serial troller' from reminting with someone elses v2 serial and simply changing the version flag to v1 when they spend the zerocoin, which would negate any of the additional v2 requirements.

It is unlikely that many (if any) serials were minted with the first byte as 00. Note that if there are existing v1 xION with 00 prefixed serials, it would make them unspendable since the owner would not have an associated private key.